improved candidate finding

+++ pepper_variant/modules/python/DataStorePredict.py
@@ -67,12 +67,10 @@ class DataStore(object):
         if name not in self.meta['predictions']:
             self.meta['predictions'].add(name)
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-                                                      chunk_name_suffix, 'position')] = np.array(position, dtype=np.uint32)
+                                                      chunk_name_suffix, 'position')] = np.array(position, dtype=np.int32)
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-                                                      chunk_name_suffix, 'index')] = np.array(index, dtype=np.uint16)
-            # self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-            #                                           chunk_name_suffix, 'ref_seq')] = ref_seq
+                                                      chunk_name_suffix, 'index')] = np.array(index, dtype=np.int32)
             self.file_handler['{}/{}/{}/{}/{}'.format(self._prediction_path_, contig, chunk_name_prefix,
-                                                      chunk_name_suffix, 'base_predictions')] = np.array(base_predictions, dtype=np.uint8)
+                                                      chunk_name_suffix, 'base_predictions')] = np.array(base_predictions, dtype=np.int32)
